requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
excel-util==0.0.4
idna==3.4
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.5.2
pandas-util==0.0.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tweepy==4.12.1
urllib3==1.26.12
google-cloud-secret-manager==1.0.0
google-cloud-bigquery